requirements.txt
schedule==1.2.1
python-dotenv==1.0.0 